requirements.txt
annotated-types
anyio
asgiref
certifi
charset-normalizer
whitenoise
click
cohere
colorama
debugpy
dj-database-url
Django
django-widget-tweaks
djangorestframework
django-environ
fastavro
filelock
fsspec
gunicorn
h11
hf-xet
httpcore
httpx
httpx-sse
huggingface-hub
idna
isort
itsdangerous
jedi
packaging
parso
pillow
platformdirs
prompt_toolkit
psycopg2-binary
pydantic
pydantic_core
python-dotenv
PyYAML
requests
shellingham
sniffio
sqlparse
tokenizers
tqdm
traitlets
typer-slim
types-requests
typing-inspection
typing_extensions
tzdata
urllib3
wcwidth
django-htmlmin